Clarify scope and purpose of Delete action
The action to remove files from the archive uses the word "Delete",
which is a scary word and does not clarify its scope at all; it could
conceivably delete the whole archive, or delete selected files from
outside the archive. In fact all the action does is remove the
selected files from the archive; the user's files outside the archive
are safe.

This commit renames the action to communicate its scope and purpose
more clearly.