cmake: add function to parse the mimetypes from plugins metadata
cmake >= 3.19 is able to parse JSON files. We can use this feature to
simplify the handling of mimetypes in the different plugins.

This commit introduces a cmake function that parses the MimeTypes field
from the JSON metadata file of a plugin into a cmake list of strings.

We need cmake >= 3.19 for the JSON parsing capability.