[libarchiveplugin] Speedup extraction
Speedup extraction large files and sparse files in LibarchivePlugin using
libarchive's zero-copy architecture instead of buffer copying.

https://github.com/libarchive/libarchive/wiki/ZeroCopy

For the ReadWriteLibarchivePlugin leave the old way of copying through
the buffer, because it opens the handle for writing using
archive_write_new() instead of archive_write_disk_new(), and
ahearchive_write_data_block() interface is not available for it.