Fix support for bzip2 format with shared-mime-info 2.3
shared-mime-info 2.3 explicitly separated the mimetypes for bzip2 archives
from the mimetypes for the (old and deprecated) bzip archives.

libarchive doesn't support the old bzip format, however we can't just
drop it from ark since we still need it on systems that use an old
shared-mime-info package.

So for now we drop it only when we are sure that we are using a
shared-mime-info >= 2.3