Fix opening of AppImage files
This hack in the mimetype detection logic in Ark was added to work
around an old bug in shared-mime-info about ISO files detected as text
files.

It should not be needed anymore nowadays, since even Debian 11 ships
shared-mime-info 2.0 which should already contain the fix.

This check currently even breaks the detection of
application/x-iso9660-appimage (since it inherits the ISO mimetype),
so it's about time we just drop it.