Fix validation of tar mimetypes
The validation code was ignoring any alpha chars *after* the extension,
which might be a pretty common case (e.g. one copies a foo.tar.gz file
into a foo.tar.gz_backup file or whatever).

It's not worth to fix this code, as it already has enough hacks added
over time.

The new approach introduced with this commit is much simpler: we just look
for the longest match between the filename's complete suffix and the
suffixes supported by the libarchive plugin. If we found a match, we
just replace it in place of the original complete suffix.

If the original suffix doesn't match, we try its stripped down version
(with non-alpha chars removed), as the old code did (see the CCBUG).

A bunch of test cases have been added (they did not pass before this
commit).

CCBUG: 328815