cliinterface: drop unnecessary qRegisterMetaType
And drop the deprecated QMetaType::type() call as well.

This was probably needed back when the CliInterface lived in another
thread, but that is no longer true since a537c5ee4a1b52d88c8d2427296534173311e474

For our needs it's enough that QProcess::ExitStatus is defined as
Q_ENUM by QProcess itself.